+++ sdk/keyvault/Azure.Security.KeyVault.Secrets/src/SecretClientOptions.cs
@@ -18,7 +18,7 @@ namespace Azure.Security.KeyVault.Secrets
         /// For more information, see
         /// <see href="https://docs.microsoft.com/rest/api/keyvault/key-vault-versions">Key Vault versions</see>.
         /// </summary>
-        internal const ServiceVersion LatestVersion = ServiceVersion.V2026_03_01_Preview;
+        internal const ServiceVersion LatestVersion = ServiceVersion.V2025_07_01;
 
         /// <summary>
         /// The versions of Azure Key Vault supported by this client library.
@@ -65,11 +65,6 @@ namespace Azure.Security.KeyVault.Secrets
             /// The Key Vault API version 2025-07-01.
             /// </summary>
             V2025_07_01 = 7,
-
-            /// <summary>
-            /// The Key Vault API version 2026-03-01-preview.
-            /// </summary>
-            V2026_03_01_Preview = 8,
 #pragma warning restore CA1707 // Identifiers should not contain underscores
         }
 
@@ -125,7 +120,6 @@ namespace Azure.Security.KeyVault.Secrets
                 ServiceVersion.V7_5 => "7.5",
                 ServiceVersion.V7_6 => "7.6",
                 ServiceVersion.V2025_07_01 => "2025-07-01",
-                ServiceVersion.V2026_03_01_Preview => "2026-03-01-preview",
                 _ => throw new ArgumentException(Version.ToString()),
             };
         }